backend/fastAPI_imageRecognition/requirements.txt
fastapi==0.115.8
uvicorn[standard]
tensorflow
pillow
numpy
requests
